requirement.txt
absl-py==2.0.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1==2.4.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bottlenose==1.1.8
cachetools==5.3.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clr-loader==0.2.6
comm==0.1.3
contourpy==1.1.1
cryptography==35.0.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-json-widget==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
enum-compat==0.0.3
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.12.2
future==0.18.3
gast==0.5.4
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.2
imageio==2.33.1
ipykernel==6.22.0
ipython==8.13.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mecab-python3==1.0.6
mercari==2.0.1
mistune==2.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
ntlm-auth==1.5.0
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.5.0
plotly==5.18.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pycparser==2.20
pygame==2.5.2
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-amazon-simple-product-api==2.2.11
python-dateutil==2.8.2
python-json-logger==2.0.7
pythonnet==3.0.3
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qiskit==0.46.0
qiskit-aer==0.13.3
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.46.0
qiskit-textbook @ git+https://github.com/qiskit-community/qiskit-textbook.git@0ae905083ec3872d7ddc6bb4001d5468f476d93f#subdirectory=qiskit-textbook-src
requests==2.26.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
rustworkx==0.14.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.13.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.1
stevedore==5.1.0
SudachiDict-core==20230110
SudachiPy==0.6.7
symengine==0.11.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
unidic-lite==1.0.8
uri-template==1.2.0
urllib3==1.26.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xlrd==2.0.1
